+++ server/routes/events.ts
@@ -24,8 +24,13 @@ import {
 } from "../types";
 //Logger
 import { logger } from "../loggerUtils/logger";
+import dotenv from "dotenv";
 
 const router = express.Router();
+dotenv.config();
+
+const TODAY =
+  process.env.NODE_ENV !== "dev" ? "TODAY()" : `"${process.env.TODAY}"`;
 
 function processGeneralEventData(event: Record<Event>): ProcessedEvent {
   const optionsDay = {
@@ -134,7 +139,7 @@ router.route("/api/events").get(
     const url =
       `${AIRTABLE_URL_BASE}/🚛 Supplier Pickup Events?` +
       // Get all events that are after yesterday. We want it to be after yesterday because we want the event to still show up on the day of the event
-      `&filterByFormula=IS_AFTER({Start Time}, (DATEADD(TODAY(), -1, 'days')))` +
+      `&filterByFormula=IS_AFTER({Start Time}, (DATEADD(${TODAY}, -1, 'days')))` +
       // Get fields for upcoming events dashboard
       `&fields=Start Time` + // Day, Time
       `&fields=Pickup Address` + // Main Location